Maze2DArray
Milestone 1: The code will scan in a file and then find the amount of rows and columns in the file. It will then put in values at each index of a declared 2d int array MazeBoard so the file matches the 2d int array. The array will then be printed out and be the same as the file chosen.

MileStone 2: The code will now recursively scan through the maze to find all path(s) and change each element of the path to '2'. The code will print out the end coordinate of the path and if there isn't an end the code will print "no end found".

Final Submission: The code will now backtrack through the maze to find the SHORTEST path int the maze. The maze will then be printed out showing this shortest path apart from the rest of the maze. This new array will then be printed into the .txt file SolvedMaze.