Add files via upload
Write a program that reads in an integer from the user and then outputs the corresponding Fibonacci number. The Fibonacci numbers are a sequence of number Fn where Fn = Fn-1 + Fn-2 and F0 is 0 and F1 is 1. You should compute your Fibonacci numbers iteratively. You should have a function that takes a number and returns its Fibonacci number.
Example Output
Enter a number: 9 F9 is 34