Add files via upload
Write a program that reads in a number of integers, then prints out the average of the values. The program should finish reading in values when it encounters the end of the file. If no values are read in from the input the program should print No numbers were read.